+++ common/serializers/stream_serializer.py
